初步利用elastic search 实现课程搜索的基本功能 todo:  价格区间的范围搜索  排序功能实现

+++ xc-framework-model/src/main/java/com/xuecheng/framework/domain/course/CoursePub.java
@@ -8,13 +8,10 @@ import javax.persistence.*;
 import java.io.Serializable;
 import java.util.Date;
 
-/**
- * Created by admin on 2018/2/10.
- */
 @Data
 @ToString
 @Entity
-@Table(name="course_pub")
+@Table(name = "course_pub")
 @GenericGenerator(name = "jpa-assigned", strategy = "assigned")
 public class CoursePub implements Serializable {
     private static final long serialVersionUID = -916357110051689487L;
@@ -35,10 +32,10 @@ public class CoursePub implements Serializable {
     private String charge;
     private String valid;
     private String qq;
-    private Float price;
-    private Float price_old;
+    private Double price;
+    private Double  price_old;
     private String expires;
     private String teachplan;//课程计划
-    @Column(name="pub_time")
+    @Column(name = "pub_time")
     private String pubTime;//课程发布时间
 }
